# Remove Chunkify method